// src/widgets/BhandaraCheckinV0/types.ts
export type UserDetailsValueWrapper<T> = {
  show?: boolean;
  isValid?: boolean;
  value?: T;
  disabled?: boolean;
};

export type UserDetails = {
  fullName: UserDetailsValueWrapper<string>;
  mobile: UserDetailsValueWrapper<string>;
  email: UserDetailsValueWrapper<string>;
  location: UserDetailsValueWrapper<RefinedCityStateCountryLocation>;
  ageGroup: UserDetailsValueWrapper<string>;
  gender: UserDetailsValueWrapper<string>;
};